feat: migrate all examples to latest schema version (v14)
Fixed migration infrastructure to enable headless migrations:
- Created render-settings-constants.ts with no dependencies
- Updated serialization.ts to re-export from constants
- Updated migration 012 to import from constants (breaks UI dependency chain)
- Migrated all 17 examples from v7 → v14

This completes migrations 008-014:
- 008: Rename DateOp to DatetimeOp
- 009: Move editor settings to project JSON
- 010: Change parent references to use IDs
- 011: Rename ForLoop 'd' field to 'item'
- 012: Move render settings to OutOp
- 013: Add time marker support to timeline
- 014: Remove MapStyleOp (inline values)
