+++ noodles-editor/src/examples/aggregation-example/noodles.json
@@ -97,23 +97,12 @@
     "y": 907.0984953781293,
     "zoom": 0.8019816514742112
   },
-  "version": 7,
-  "timeline": {
-    "sheetsById": {
-      "Noodles": {
-        "staticOverrides": {
-          "byObject": {}
-        }
-      }
-    },
-    "definitionVersion": "0.4.0",
-    "revisionHistory": [
-      "fBu1FW4HYanu9xRt",
-      "aiRAn66WU7kCXJZh",
-      "mtj8EgxoAJHG9vWu"
-    ]
-  },
+  "version": 14,
   "name": "SF Bike Parking Aggregation",
+  "editorSettings": {
+    "layoutMode": "noodles-on-top",
+    "showOverlay": true
+  },
   "nodes": [
     {
       "id": "/deck",
@@ -131,7 +120,19 @@
       "id": "/out",
       "type": "OutOp",
       "data": {
-        "inputs": {},
+        "inputs": {
+          "display": "fixed",
+          "width": 1920,
+          "height": 1080,
+          "lod": 2,
+          "waitForData": true,
+          "codec": "avc",
+          "bitrateMbps": 10,
+          "bitrateMode": "constant",
+          "scaleControl": 0.3,
+          "framerate": 30,
+          "captureDelay": 50
+        },
         "locked": false
       },
       "position": {
@@ -245,5 +246,20 @@
         "locked": false
       }
     }
-  ]
+  ],
+  "timeline": {
+    "sheetsById": {
+      "Noodles": {
+        "staticOverrides": {
+          "byObject": {}
+        }
+      }
+    },
+    "definitionVersion": "0.4.0",
+    "revisionHistory": [
+      "fBu1FW4HYanu9xRt",
+      "aiRAn66WU7kCXJZh",
+      "mtj8EgxoAJHG9vWu"
+    ]
+  }
 }
