+++ noodles-editor/src/noodles/__migrations__/012-render-settings-to-outop.ts
@@ -1,5 +1,6 @@
-import type { NoodlesProjectJSON, RenderSettings } from '../utils/serialization'
-import { DEFAULT_RENDER_SETTINGS } from '../utils/serialization'
+import type { NoodlesProjectJSON } from '../utils/serialization'
+import type { RenderSettings } from '../utils/render-settings-constants'
+import { DEFAULT_RENDER_SETTINGS } from '../utils/render-settings-constants'
 
 // Migration to move render settings from Theatre.js staticOverrides and project-level settings
 // to OutOp input fields.
