+++ noodles-editor/src/noodles/utils/serialization.ts
@@ -19,33 +19,8 @@ export type EditorSettings = {
   showDebugInfo?: boolean
 }
 
-export type RenderSettings = {
-  display: 'fixed' | 'responsive'
-  resolution: { width: number; height: number }
-  lod: number
-  waitForData: boolean
-  codec: 'avc' | 'hevc' | 'vp9' | 'av1'
-  bitrateMbps: number
-  bitrateMode: 'constant' | 'variable'
-  scaleControl: number
-  framerate: number
-  captureDelay: number
-  rendersDirectory: string
-}
-
-export const DEFAULT_RENDER_SETTINGS: RenderSettings = {
-  display: 'fixed',
-  resolution: { width: 1920, height: 1080 },
-  lod: 2,
-  waitForData: true,
-  codec: 'avc',
-  bitrateMbps: 10,
-  bitrateMode: 'constant',
-  scaleControl: 0.3,
-  framerate: 30,
-  captureDelay: 50,
-  rendersDirectory: 'renders',
-}
+export type { RenderSettings } from './render-settings-constants'
+export { DEFAULT_RENDER_SETTINGS } from './render-settings-constants'
 
 export type NoodlesProjectJSON = ReactFlowJsonObject & {
   version: number
